check for sites-folder and only do actions if it exists

+++ wema.ps1
@@ -34,13 +34,15 @@ if ($command -eq "fetch") {
   Remove-Item -Recurse -Force .\web\user\config
   Move-Item -Path .\DATA_REPO_TEMP\config -Destination .\web\user
   # Loop over folders in .\DATA_REPO_TEMP\sites
-  $sites = Get-ChildItem -Path .\DATA_REPO_TEMP\sites -Directory
-  foreach ($site in $sites) {
-    $siteName = $site.Name
-    Remove-Item -Recurse -Force .\web\user\sites\$siteName\config
-    Remove-Item -Recurse -Force .\web\user\sites\$siteName\pages
-    Move-Item -Path .\DATA_REPO_TEMP\sites\$siteName\config -Destination .\web\user\sites\$siteName
-    Move-Item -Path .\DATA_REPO_TEMP\sites\$siteName\pages -Destination .\web\user\sites\$siteName
+  if (Test-Path -Path .\DATA_REPO_TEMP\sites) {
+    $sites = Get-ChildItem -Path .\DATA_REPO_TEMP\sites -Directory
+    foreach ($site in $sites) {
+	  $siteName = $site.Name
+	  Remove-Item -Recurse -Force .\web\user\sites\$siteName\config
+	  Remove-Item -Recurse -Force .\web\user\sites\$siteName\pages
+	  Move-Item -Path .\DATA_REPO_TEMP\sites\$siteName\config -Destination .\web\user\sites\$siteName
+	  Move-Item -Path .\DATA_REPO_TEMP\sites\$siteName\pages -Destination .\web\user\sites\$siteName
+    }
   }
 
   Remove-Item -Recurse -Force DATA_REPO_TEMP
